pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "issues-agent"
version = "0.1.0"
description = "GitHub Issues AI Agent for classification and reporting"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Your Name", email = "you@example.com" },
]
dependencies = [
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "pyyaml>=6.0",
]

[tool.setuptools.packages.find]
include = ["issues_agent"]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
]
